♻️ refactor: Separate login toast logic as custom hook & apply global Toaster - Create and apply useLoginToast custom hook - Refactor LoginToast component to only call the hook - Place Toaster globally in layout.tsx for toast support on all pages - Show sonner toast on logout from Header - Follow client/server component rules and organize related file paths Ref #51
